--- modules/activiti-spring-data-jpa/src/main/java/org/activiti/springdatajpa/pojo/ActHiIdentitylink.java
@@ -1,104 +1,0 @@
-package org.activiti.springdatajpa.pojo;
-// Generated Nov 21, 2015 11:41:58 AM by Hibernate Tools 3.2.2.GA
-
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-/**
- * ActHiIdentitylink generated by hbm2java
- */
-@Entity
-@Table(name="act_hi_identitylink"
-    ,schema="public"
-)
-public class ActHiIdentitylink  implements java.io.Serializable {
-
-
-     private String id;
-     private String groupId;
-     private String type;
-     private String userId;
-     private String taskId;
-     private String procInstId;
-
-    public ActHiIdentitylink() {
-    }
-
-	
-    public ActHiIdentitylink(String id) {
-        this.id = id;
-    }
-    public ActHiIdentitylink(String id, String groupId, String type, String userId, String taskId, String procInstId) {
-       this.id = id;
-       this.groupId = groupId;
-       this.type = type;
-       this.userId = userId;
-       this.taskId = taskId;
-       this.procInstId = procInstId;
-    }
-   
-     @Id 
-    
-    @Column(name="id_", unique=true, nullable=false, length=64)
-    public String getId() {
-        return this.id;
-    }
-    
-    public void setId(String id) {
-        this.id = id;
-    }
-    
-    @Column(name="group_id_")
-    public String getGroupId() {
-        return this.groupId;
-    }
-    
-    public void setGroupId(String groupId) {
-        this.groupId = groupId;
-    }
-    
-    @Column(name="type_")
-    public String getType() {
-        return this.type;
-    }
-    
-    public void setType(String type) {
-        this.type = type;
-    }
-    
-    @Column(name="user_id_")
-    public String getUserId() {
-        return this.userId;
-    }
-    
-    public void setUserId(String userId) {
-        this.userId = userId;
-    }
-    
-    @Column(name="task_id_", length=64)
-    public String getTaskId() {
-        return this.taskId;
-    }
-    
-    public void setTaskId(String taskId) {
-        this.taskId = taskId;
-    }
-    
-    @Column(name="proc_inst_id_", length=64)
-    public String getProcInstId() {
-        return this.procInstId;
-    }
-    
-    public void setProcInstId(String procInstId) {
-        this.procInstId = procInstId;
-    }
-
-
-
-
-}
-
-
