+++ modules/activiti-spring-data-jpa/src/main/java/org/activiti/springdatajpa/pojo/HistoricIdentityLink.java
@@ -1,0 +1,96 @@
+package org.activiti.springdatajpa.pojo;
+// Generated Nov 21, 2015 11:41:58 AM by Hibernate Tools 3.2.2.GA
+
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+/**
+ * HistoricIdentityLink generated by hbm2java
+ */
+@Entity
+@Table(name = "act_hi_identitylink")
+public class HistoricIdentityLink implements java.io.Serializable {
+
+    private String id;
+    private String groupId;
+    private String type;
+    private String userId;
+    private String taskId;
+    private String proccessInstanceId;
+
+    public HistoricIdentityLink() {
+    }
+
+    public HistoricIdentityLink(String id) {
+        this.id = id;
+    }
+
+    public HistoricIdentityLink(String id, String groupId, String type, String userId, String taskId, String proccessInstanceId) {
+        this.id = id;
+        this.groupId = groupId;
+        this.type = type;
+        this.userId = userId;
+        this.taskId = taskId;
+        this.proccessInstanceId = proccessInstanceId;
+    }
+
+    @Id
+    @Column(name = "id_", unique = true, nullable = false, length = 64)
+    public String getId() {
+        return this.id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    @Column(name = "group_id_")
+    public String getGroupId() {
+        return this.groupId;
+    }
+
+    public void setGroupId(String groupId) {
+        this.groupId = groupId;
+    }
+
+    @Column(name = "type_")
+    public String getType() {
+        return this.type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    @Column(name = "user_id_")
+    public String getUserId() {
+        return this.userId;
+    }
+
+    public void setUserId(String userId) {
+        this.userId = userId;
+    }
+
+    @Column(name = "task_id_", length = 64)
+    public String getTaskId() {
+        return this.taskId;
+    }
+
+    public void setTaskId(String taskId) {
+        this.taskId = taskId;
+    }
+
+    @Column(name = "proc_inst_id_", length = 64)
+    public String getProccessInstanceId() {
+        return this.proccessInstanceId;
+    }
+
+    public void setProccessInstanceId(String proccessInstanceId) {
+        this.proccessInstanceId = proccessInstanceId;
+    }
+}
+
+
